+++ src/components/Tree/index.jsx
@@ -1,4 +1,3 @@
-import { PureComponent } from 'react'
 import TreeNodeItem from './TreeNodeItem'
 
 import './index.less'