+++ src/pages/Cms/components/components/LineVideo/config/index.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 const Component: React.FC = (): JSX.Element => {
   return <div>video</div>
 }